Ant drawing pheromones

+++ Ported/Assets/Components/Pheromone.cs
@@ -1,8 +1,9 @@
 using Unity.Entities;
 using Unity.Mathematics;
+using UnityEngine;
 
 [InternalBufferCapacity(1024)]
 public struct Pheromone : IBufferElementData
 {
-    public byte Value;
+    public float Value;
 }
